Add a simple binary file writing backend.
The binary format is similar to the text format: A valid OSC timetag,
an OSC string for the path, followed by a single character for the
type, an int of the length of the vector, and the float or int values
of the vector.  All multibyte values are written in big-endian.